python number gui property fix
git-svn-id: http://fs.math.tu-berlin.de:8000/svn/jreality/trunk@5763 6a571d9c-c4e9-67b8-d2a0-cd0d85a3e679

+++ src-plugin/de/jreality/plugin/scripting/gui/NumberSpinnerGUI.java
@@ -69,7 +69,7 @@ public class NumberSpinnerGUI extends PythonGUIPlugin<Number> {
 		}
 		@Override
 		public void setVariableValue(Number val) {
-			frontend.model.setValue(val);
+			frontend.model.setValue(val.doubleValue());
 		}
 
 		
@@ -106,9 +106,9 @@ public class NumberSpinnerGUI extends PythonGUIPlugin<Number> {
 			
 			backend.setListenersEnabled(false);
 			try {
-				backend.setMaxValue(maxValue);
-				backend.setMinValue(minValue);
-				backend.setStepSize(stepSize);
+				backend.setMaxValue(maxValue.doubleValue());
+				backend.setMinValue(minValue.doubleValue());
+				backend.setStepSize(stepSize.doubleValue());
 				backend.setIntegerValue(isInteger);
 				backend.setInstantExecute(isInstant());
 				backend.setShowSlider(showSlider);
